ya despliega los productos de la base en pantalla
